requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==2.4.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.0
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
filetype==1.2.0
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.41.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.77
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.32
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.6
narwhals==2.6.0
nest-asyncio==1.5.8
numpy==2.3.3
orjson==3.11.3
ormsgpack==1.10.0
packaging==23.2
pandas==2.3.3
parso==0.8.3
pillow==11.3.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.6
pure-eval==0.2.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.2.5
python-dateutil==2.8.2
pytz==2025.2
pywin32==306
PyYAML==6.0.3
pyzmq==25.1.1
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.3.3
tqdm==4.67.1
traitlets==5.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.12
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
